requirements.txt
python==3.7
django==3.2.0
djangorestframework